# internal/httpserver/problem.go
package httpserver

import (
	"encoding/json"
	"net/http"
)

// Problem is a minimal RFC 7807 problem detail object.
type Problem struct {
	Type   string `json:"type"`
	Title  string `json:"title"`
	Detail string `json:"detail"`
	Status int    `json:"status"`
}

func writeProblem(w http.ResponseWriter, status int, title, detail string) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(Problem{
		Type:   "about:blank",
		Title:  title,
		Detail: detail,
		Status: status,
	})
}